Moving products array to config file
