+++ config/products.ts
@@ -1,0 +1,56 @@
+import { Product } from 'common/types'
+
+const products: Product[] = [
+  {
+    name: 'Terraform',
+    slug: 'terraform',
+    url: 'https://www.terraform.io/docs/index.html',
+  },
+  {
+    name: 'Packer',
+    slug: 'packer',
+    url: 'https://www.packer.io/docs',
+  },
+  {
+    name: 'Consul',
+    slug: 'consul',
+    url: 'https://www.consul.io/docs',
+  },
+  {
+    name: 'Vault',
+    slug: 'vault',
+    url: 'https://www.vaultproject.io/docs',
+  },
+  {
+    name: 'Boundary',
+    slug: 'boundary',
+    url: 'https://www.boundaryproject.io/docs',
+  },
+  {
+    name: 'Nomad',
+    slug: 'nomad',
+    url: 'https://www.nomadproject.io/docs',
+  },
+  {
+    name: 'Waypoint',
+    slug: 'waypoint',
+    url: '/waypoint/docs',
+  },
+  {
+    name: 'Vagrant',
+    slug: 'vagrant',
+    url: 'https://www.vagrantup.com/docs',
+  },
+  {
+    name: 'Sentinel',
+    slug: 'sentinel',
+    url: 'https://docs.hashicorp.com/sentinel',
+  },
+  {
+    name: 'HashiCorp Cloud Platform',
+    slug: 'hcp',
+    url: 'https://cloud.hashicorp.com/docs/hcp',
+  },
+]
+
+export { products }
